Refactor Hackathon Detail Page Code
Move the Components to /Components folder in order to reduce number
of lines in the file and easier developer experience while making
changes.
