+++ components/HackathonDetailPage/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Leaderboard"
+export * from "./Overview"
+export * from "./Participants"
