+++ pages/hackathon/[slug]/index.tsx
@@ -1,30 +1,21 @@
-import Link from "next/link"
-import { useRouter } from "next/router"
-import { PropsWithChildren, useEffect, useState } from "react"
-import axios from "../../../util/axios"
-import { useAuth } from "../../../context/auth"
-import {
-    Nav,
-    Table,
-    Tab,
-    Container,
-    Spinner,
-    Image,
-    Badge,
-    Button,
-} from "react-bootstrap"
-import { Text } from "atomize"
-import ReactMarkdown from "react-markdown"
-import gfm from "remark-gfm"
-import highlight from "remark-highlight.js"
-import codeformatter from "remark-code-frontmatter"
 import Head from "next/head"
+import Link from "next/link"
+import { Text } from "atomize"
+import { useRouter } from "next/router"
+import { useEffect, useState } from "react"
+import { Nav, Tab, Container, Spinner, Image, Button } from "react-bootstrap"
 import {
     HackathonDetailSerializer,
     SubmissionsSerializer,
-    TeamDetailSerializer,
     TeamSerializer,
 } from "@/types/backend"
+import { useAuth } from "../../../context/auth"
+import axios from "../../../util/axios"
+import {
+    Leaderboard,
+    Overview,
+    Participants,
+} from "../../../components/HackathonDetailPage"
 
 export default function Hackathon() {
     const router = useRouter()
@@ -32,10 +23,10 @@ export default function Hackathon() {
 
     const { token, profile, loading } = useAuth()
     const [localLoading, setLocalLoading] = useState<boolean>(true)
+    const [error, setError] = useState<number>(0)
     const [hackathon, setHackathon] = useState<HackathonDetailSerializer>(
         {} as HackathonDetailSerializer
     )
-    const [error, setError] = useState<number>(0)
 
     useEffect(() => {
         if (slug) {
@@ -500,393 +491,5 @@ export default function Hackathon() {
                 </div>
             )}
         </>
-    )
-}
-
-type OverviewPropTypes = PropsWithChildren<{
-    hackathon: HackathonDetailSerializer
-}>
-function Overview({ hackathon }: OverviewPropTypes) {
-    return (
-        <div className="overview_body">
-            <div className="p-3">
-                <Text
-                    tag="h1"
-                    textSize="display1"
-                    m={{ b: "1rem" }}
-                    fontFamily="madetommy-bold"
-                >
-                    {hackathon.title}
-                </Text>
-
-                <ReactMarkdown
-                    plugins={[gfm, codeformatter, highlight]}
-                    className="pb-3 markabcs"
-                >
-                    {hackathon.description}
-                </ReactMarkdown>
-                <div className="pb-3">
-                    <Text
-                        tag="h6"
-                        textSize="subheader"
-                        fontFamily="madetommy-bold"
-                    >
-                        START DATE:
-                    </Text>
-                    {new Date(hackathon.start).toString()}
-                </div>
-                <div className="pb-3">
-                    <Text
-                        tag="h6"
-                        textSize="subheader"
-                        fontFamily="madetommy-bold"
-                    >
-                        END DATE:
-                    </Text>
-                    {new Date(hackathon.end).toString()}
-                </div>
-                {/* <div className="pb-3">
-                    <Text
-                        tag="h6"
-                        textSize="subheader"
-                        fontFamily="madetommy-bold"
-                    >
-                        STATUS:
-                    </Text>
-                    {hackathon.status}
-                </div> */}
-                <div className="pb-3">
-                    <Text
-                        tag="h6"
-                        textSize="subheader"
-                        fontFamily="madetommy-bold"
-                    >
-                        MAX TEAM SIZE:
-                    </Text>
-                    {hackathon.max_team_size}
-                </div>
-                <div className="pb-3">
-                    <Text
-                        tag="h6"
-                        textSize="subheader"
-                        fontFamily="madetommy-bold"
-                    >
-                        RESULTS DECLARED:
-                    </Text>
-                    {hackathon.results_declared ? "Yes" : "No"}
-                </div>
-            </div>
-            <style jsx>{`
-                .p3 {
-                    margin-left: 50px;
-                }
-                .overview_body {
-                    margin-left: 50px;
-                }
-                .title {
-                    font-size: 30px;
-                    padding-bottom: 30px;
-                    padding-top: 30px;
-                }
-            `}</style>
-        </div>
-    )
-}
-
-type ParticipantsPropTypes = PropsWithChildren<{ slug: string }>
-function Participants({ slug }: ParticipantsPropTypes) {
-    const [teams, setTeams] = useState<TeamDetailSerializer[]>([])
-    const [loading, setLoading] = useState<boolean>(true)
-
-    useEffect(() => {
-        setLoading(true)
-        if (slug) {
-            axios
-                .get<TeamDetailSerializer[]>(`/hackathons/${slug}/teams/`)
-                .then((response) => {
-                    setTeams(response.data)
-                })
-                .catch(() => {})
-                .finally(() => setLoading(false))
-        }
-    }, [slug])
-
-    return (
-        <div>
-            {loading ? (
-                <div className="text-center">
-                    <Table
-                        responsive
-                        style={{
-                            borderCollapse: "separate",
-                            borderSpacing: "0px 5px",
-                        }}
-                    >
-                        <thead
-                            style={{
-                                backgroundColor: "rgba(0.9,0,0,0.04)",
-                            }}
-                        >
-                            <tr>
-                                <th>Name</th>
-                                <th>Team Name</th>
-                                <th>Username</th>
-                            </tr>
-                        </thead>
-                    </Table>
-                    <Spinner animation="border" role="status">
-                        <span className="sr-only">Loading...</span>
-                    </Spinner>
-                </div>
-            ) : (
-                <div>
-                    <Table
-                        responsive
-                        style={{
-                            borderCollapse: "separate",
-                            borderSpacing: "0px 5px",
-                        }}
-                    >
-                        <thead
-                            style={{
-                                backgroundColor: "rgba(0.9,0,0,0.04)",
-                            }}
-                        >
-                            <tr>
-                                <th>Name</th>
-                                <th>Team Name</th>
-                                <th>Username</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {teams.map((team) => {
-                                var members = []
-                                members.push(
-                                    <tr
-                                        key={team.name + "_" + team.leader.name}
-                                        className="bg-grey rounded"
-                                    >
-                                        <td>
-                                            {team.leader.name}{" "}
-                                            <Badge variant="info">Leader</Badge>
-                                        </td>
-                                        <td>{team.name}</td>
-                                        <td>
-                                            <Link
-                                                href={`/profile/${team.leader.username}`}
-                                                passHref
-                                            >
-                                                <a href="/#">
-                                                    {team.leader.username}
-                                                </a>
-                                            </Link>
-                                        </td>
-                                    </tr>
-                                )
-                                for (var i = 0; i < team.members.length; i++) {
-                                    if (
-                                        team.members[i].username !=
-                                        team.leader.username
-                                    ) {
-                                        members.push(
-                                            <tr
-                                                key={
-                                                    team.name +
-                                                    "_" +
-                                                    team.members[i].username
-                                                }
-                                                className="bg-grey rounded"
-                                            >
-                                                <td>{team.members[i].name}</td>
-                                                <td>{team.name}</td>
-                                                <td>
-                                                    <Link
-                                                        href={`/profile/${team.members[i].username}`}
-                                                        passHref
-                                                    >
-                                                        <a href="/#">
-                                                            {
-                                                                team.members[i]
-                                                                    .username
-                                                            }
-                                                        </a>
-                                                    </Link>
-                                                </td>
-                                            </tr>
-                                        )
-                                    }
-                                }
-                                return members
-                            })}
-                        </tbody>
-                    </Table>
-                    {!teams.length ? (
-                        <div className="bg-grey text-center py-3 rounded">
-                            No Participants Found
-                        </div>
-                    ) : null}
-                    <style jsx>{`
-                        .bg-grey {
-                            margin-top: 2px;
-                            background-color: rgba(0.9, 0, 0, 0.04);
-                            height: 7vh;
-                            margin-bottom: 4px;
-                        }
-                    `}</style>
-                </div>
-            )}
-        </div>
-    )
-}
-
-type LeaderboardPropTypes = PropsWithChildren<{
-    status: string
-    submissions: SubmissionsSerializer[]
-    loading: boolean
-}>
-function Leaderboard({ status, submissions, loading }: LeaderboardPropTypes) {
-    const submissionList = submissions
-        .sort((a, b) => b.score - a.score)
-        .map((submission, index) => (
-            <tr key={index}>
-                <td>{index + 1}</td>
-                <td>{submission.teamName} </td>
-                <td> {submission.score}</td>
-                <td>
-                    <Link href={`/submission/${submission.id}`} passHref>
-                        <a href="/#">{submission.title}</a>
-                    </Link>
-                </td>
-            </tr>
-        ))
-    return (
-        <div>
-            <Table responsive>
-                <thead
-                    style={{
-                        backgroundColor: "rgba(0.9,0,0,0.04)",
-                    }}
-                >
-                    <tr>
-                        <th>Rank</th>
-                        <th>Team Name</th>
-                        <th>Score</th>
-                        <th>Submission Link</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {status == "Upcoming" ? (
-                        <tr>
-                            <td colSpan={4} className="px-3">
-                                <div
-                                    className="row rounded-bottom"
-                                    style={{
-                                        backgroundColor: "rgba(0.9,0,0,0.04)",
-                                    }}
-                                >
-                                    <div className="col-12 text-center py-3">
-                                        <Image
-                                            src="/images/rocket.svg"
-                                            style={{ height: "80px" }}
-                                            alt="Rocket Icon"
-                                        />
-                                        <p>Let the Hackathon Begin</p>
-                                    </div>
-                                </div>
-                            </td>
-                        </tr>
-                    ) : status == "Ongoing" ? (
-                        <tr>
-                            <td colSpan={4} className="px-3">
-                                <div
-                                    className="row rounded-bottom"
-                                    style={{
-                                        backgroundColor: "rgba(0.9,0,0,0.04)",
-                                    }}
-                                >
-                                    <div className="col-12 text-center py-3">
-                                        <Image
-                                            src="/images/rocket.svg"
-                                            style={{ height: "80px" }}
-                                            alt="Rocket Icon"
-                                        />
-                                        <p>
-                                            Submissions can only be viewed once
-                                            the hackathon has ended!
-                                        </p>
-                                    </div>
-                                </div>
-                            </td>
-                        </tr>
-                    ) : (
-                        <>
-                            {loading ? (
-                                <tr>
-                                    <td colSpan={4}>
-                                        <Spinner
-                                            style={{
-                                                position: "absolute",
-                                                left: "50%",
-                                            }}
-                                            animation="border"
-                                            role="status"
-                                        >
-                                            <span className="sr-only">
-                                                Loading...
-                                            </span>
-                                        </Spinner>
-                                    </td>
-                                </tr>
-                            ) : (
-                                <>
-                                    {submissionList.length ? (
-                                        <>{submissionList}</>
-                                    ) : (
-                                        <tr>
-                                            <td colSpan={4} className="px-3">
-                                                <div
-                                                    className="row rounded-bottom"
-                                                    style={{
-                                                        backgroundColor:
-                                                            "rgba(0.9,0,0,0.04)",
-                                                    }}
-                                                >
-                                                    <div className="col-12 text-center py-3">
-                                                        <Text textSize="display1">
-                                                            <span
-                                                                role="img"
-                                                                aria-label="sad face"
-                                                            >
-                                                                😔
-                                                            </span>
-                                                        </Text>
-                                                        <Text>
-                                                            No Submissions Found
-                                                        </Text>
-                                                    </div>
-                                                </div>
-                                            </td>
-                                        </tr>
-                                    )}
-                                </>
-                            )}
-                        </>
-                    )}
-                </tbody>
-            </Table>
-            <style jsx>{`
-                .bg-grey {
-                    background-color: rgba(0.9, 0, 0, 0.04);
-                    display: flex;
-                    height: 7vh;
-                    margin-bottom: 4px;
-                    color: rgba(0.9, 0, 0, 0.8);
-                    padding: 10px;
-                }
-                .heading {
-                    color: black;
-                }
-            `}</style>
-        </div>
     )
 }
